backend/requirements.txt
blinker==1.9.0
chardet==5.2.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pillow==11.1.0
pycryptodome==3.21.0
qrcode==8.0
reportlab==4.3.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
